Combine tracerAPI calls into one folder
